Extract out tracing
